requirements.txt
selenium
bs4
requests
TikTokApi
Flask